Refactor tracksyscall to be able to monitor runcommand
