+++ bb/src/CMakeLists.txt
@@ -32,9 +32,9 @@ add_custom_command(
         )
 add_custom_target(need_bbras ALL DEPENDS ${CMAKE_CURRENT_BINARY_DIR}/bbras.h)
 
-add_executable(bbServer tracksyscall.cc bbconndata.cc main.cc connections.cc bberror.cc bbinternal.cc bbserver.cc xfer.cc usage.cc fh.cc serial.cc bbio.cc bbio_regular.cc bbio_BSCFS.cc BBTransferDef.cc BBJob.cc BBTagID.cc BBTagParts.cc BBTagInfo.cc BBTagInfoMap.cc BBLV_ExtentInfo.cc BBLV_Info.cc BBLV_Metadata.cc nodecontroller.cc bbwrkqmgr.cc bbwrkqe.cc ContribFile.cc ContribIdFile.cc HandleFile.cc LVUuidFile.cc weak.cc)
+add_executable(bbServer tracksyscallMonitor.cc bbconndata.cc main.cc connections.cc bberror.cc bbinternal.cc bbserver.cc xfer.cc usage.cc fh.cc serial.cc bbio.cc bbio_regular.cc bbio_BSCFS.cc BBTransferDef.cc BBJob.cc BBTagID.cc BBTagParts.cc BBTagInfo.cc BBTagInfoMap.cc BBLV_ExtentInfo.cc BBLV_Info.cc BBLV_Metadata.cc nodecontroller.cc bbwrkqmgr.cc bbwrkqe.cc ContribFile.cc ContribIdFile.cc HandleFile.cc LVUuidFile.cc weak.cc)
 
-add_executable(bbProxy tracksyscall.cc bbconndata.cc main.cc connections.cc bbproxyConn2bbserver.cc bberror.cc bbinternal.cc bbproxy.cc lvlookup.cc fh.cc serial.cc usage.cc BBTransferDef.cc BBJob.cc nodecontroller.cc LVUtils.cc weak.cc bbGrabStderr.cc ${SRCCSM})
+add_executable(bbProxy tracksyscallMonitor.cc bbconndata.cc main.cc connections.cc bbproxyConn2bbserver.cc bberror.cc bbinternal.cc bbproxy.cc lvlookup.cc fh.cc serial.cc usage.cc BBTransferDef.cc BBJob.cc nodecontroller.cc LVUtils.cc weak.cc bbGrabStderr.cc ${SRCCSM})
 
 add_library(bbAPI SHARED  bbconndata.cc connections.cc bberror.cc bbinternal.cc bbapi.cc bbapi2.cc fh.cc BBTransferDef.cc BBJob.cc nodecontroller.cc bbGrabStderr.cc  ${SRCCSM})
 add_executable(bbcmd bbcmd.cc)
