+++ utilities/src/util.cc
@@ -38,6 +38,7 @@
 #include "util.h"
 #include "logging.h"
 #include "tstate.h"
+#include "tracksyscall.h"
 
 namespace bfs = boost::filesystem;
 
@@ -374,12 +375,16 @@ std::vector<std::string> runCommand(const std::string& cmd, bool flatfile,bool n
     if(flatfile)
     {
         LOG(bb,always) << "Reading: " << cmd;
+        threadLocalTrackSyscallPtr->nowTrack(TrackSyscall::Runcommandfopen, cmd.c_str(), __LINE__);
         f = fopen(cmd.c_str(), "r");
+        threadLocalTrackSyscallPtr->clearTrack();
     }
     else
     {
         LOG(bb,always) << "Executing: " << cmd;
+        threadLocalTrackSyscallPtr->nowTrack(TrackSyscall::Runcommandpopen, cmd.c_str(), __LINE__);
         f = popen(cmd.c_str(), "r");
+        threadLocalTrackSyscallPtr->clearTrack();
     }
 
     if (f)
